Change Cell text proptypes to Text#PropTypes
It doesn't make sense to reuse View's style PropTypes for the structure of
the styles applied to the `title` and `detail` props. (For instance, React
complains when I set a fontWeight property on them, since that's only on
Text, not View.)

This PR changes them from View#propTypes to Text#propTypes.